requirements.txt
django>=5.2
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
django-celery-beat>=2.6
djangorestframework>=3.16
djangorestframework-simplejwt>=5.5
PyJWT>=2.10
